Dropped hyperglot dependency due to licensing.
 - Removed SupportedLanguages method as it relies on hyperglot, which is under the GPLv3, to keep gftranslate under Apache 2.0
 - The method is still available as a separate code-snippet (not part of gflanguages itself), and any program using that snippet will need to comply with the GPLv3.
 - For more details, see: https://github.com/googlefonts/fontbakery/pull/3617#issuecomment-1044898812
